% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div.R
\name{div}
\alias{div}
\title{División entre dos números}
\usage{
div(a, b)
}
\arguments{
\item{a}{Primer número}

\item{b}{Segundo númeron no nulo}
}
\value{
La división entre a y b
}
\description{
Esta función devuelve la división de dos números
}
\examples{
div(6,3)

div(20,10)

div(2.5,0.5)

div(1+2,6)

}
\seealso{
\link{mult}
}
