% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult.R
\name{mult}
\alias{mult}
\title{Multiplicación entre dos números}
\usage{
mult(a, b)
}
\arguments{
\item{a}{Primer número}

\item{b}{Segundo número}
}
\value{
El producto entre a y b
}
\description{
Esta función devuelve el producto de dos números
}
\examples{
mult(5,3)

}
\seealso{
\link{div}
}
