% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resta.R
\name{resta}
\alias{resta}
\title{Resta entre dos números}
\usage{
resta(a, b)
}
\arguments{
\item{a}{Primer número}

\item{b}{Segundo número}
}
\value{
La resta entre a y b
}
\description{
Esta función devuelve la resta de dos números
}
\examples{
resta(5,3)

}
\seealso{
\link{suma}
}
